Dependency injection at module level to keep dependency management clean and readable
