+++ data/src/main/java/io/chicago/data/api/ArtWorkApiService.kt
@@ -7,6 +7,7 @@ import retrofit2.http.Query
 const val baseUrl = "https://api.artic.edu/api/v1/"
 
 interface ArtWorkApiService {
+
     @GET("artworks")
     suspend fun getArtworks(@Query("limit") limit: Int = 25): ArtWorks
 }
