+++ data/src/main/java/io/chicago/data/domainImplementations/usecases/GetArtWorkUseCaseImpl.kt
@@ -1,0 +1,8 @@
+package io.chicago.data.domainImplementations.usecases
+
+import io.chicago.domain.repositories.ArtWorkRepository
+import io.chicago.domain.usecases.GetArtWorkUseCase
+
+class GetArtWorkUseCaseImpl(private val artWorkRepository: ArtWorkRepository) : GetArtWorkUseCase {
+    override suspend operator fun invoke() = artWorkRepository.fetchArtworks()
+}
