+++ domain/build.gradle
@@ -13,7 +13,9 @@ java {
 }
 
 dependencies {
+
     implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.5.0'
+
     // Coroutines
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0'
 }
