+++ domain/src/main/java/io/chicago/domain/usecases/GetArtWorkUseCase.kt
@@ -1,7 +1,8 @@
 package io.chicago.domain.usecases
 
-import io.chicago.domain.repositories.ArtWorkRepository
+import io.chicago.domain.common.DefaultViewState
+import io.chicago.domain.entities.ArtWorks
 
-class GetArtWorkUseCase(private val artWorkRepository: ArtWorkRepository) {
-    suspend operator fun invoke() = artWorkRepository.fetchArtworks()
+interface GetArtWorkUseCase {
+    suspend operator fun invoke(): DefaultViewState<ArtWorks>
 }
